requirements.txt
# API y AI
anthropic>=0.40.0
openai>=1.50.0

# Telegram Bot
python-telegram-bot==21.5

# Procesamiento de documentos
PyPDF2==3.0.1
pdfplumber==0.11.4

# Vector Store y Embeddings
chromadb>=0.5.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

# Servidor y APIs
fastapi>=0.100.0
uvicorn>=0.30.0
pydantic>=2.0.0

# Utilidades
python-dotenv>=1.0.0
requests>=2.30.0
numpy>=1.24.0
pandas>=2.0.0

# Logging y monitoreo
loguru>=0.7.0

# HTTP clients compatibles
httpx>=0.25.0

# Testing (opcional)
pytest>=8.0.0
pytest-asyncio>=0.24.0

fastapi>=0.100.0
uvicorn>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.30.0
loguru>=0.7.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
